Introduce ma.utils.conf
