+++ tests/ma/utils/test_conf.py
@@ -1,0 +1,7 @@
+from ma.utils.conf import get_code_version
+
+
+def test_get_code_version() -> None:
+    version = get_code_version()
+    assert version is not None
+    assert version != "unknown"
